fix unbound var in cleaner

+++ helpers/helpers_emba_helpers.sh
@@ -92,7 +92,7 @@ cleaner() {
     print_output "[*] Interrupt detected!" "no_log"
   fi
   print_output "[*] Final cleanup started." "no_log"
-  if [[ -n "${QUEST_CONTAINER}" ]]; then
+  if [[ "$IN_DOCKER" -eq 0 ]] && [[ -n "${QUEST_CONTAINER}" ]]; then
     if [[ "$( docker container inspect -f '{{.State.Status}}' "${QUEST_CONTAINER}" )" == "running" ]]; then
       print_output "[*] Stopping Quest Container ..." "no_log"
       docker kill "${QUEST_CONTAINER}"
